Refactor: 리퀘스트
