+++ app/src/main/java/com/seungma/infratalk/data/model/request/user/SignupRequest.kt
@@ -1,4 +1,4 @@
-package com.seungma.infratalk.data.model.request
+package com.seungma.infratalk.data.model.request.user
 
 data class SignupRequest(
     val email: String,
